Move fetch size estimator to upstream module
