+++ airbyte-db/lib/src/main/java/io/airbyte/db/jdbc/FetchSizeEstimator.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.relationaldb;
+package io.airbyte.db.jdbc;
 
 import com.google.common.annotations.VisibleForTesting;
